+++ tensorflow_datasets/image/duke_ultrasound_test.py
@@ -10,19 +10,26 @@ from tensorflow_datasets.image import duke_ultrasound
 
 class DukeUltrasoundTest(testing.DatasetBuilderTestCase):
   DATASET_CLASS = duke_ultrasound.DukeUltrasound
+  OVERLAPPING_SPLITS = ['A', 'B', 'TRAIN']
+
   SPLITS = {
     "train": 1,
     "test": 1,
     "validation": 1,
-    "invivo": 1
+    "MARK": 1,
+    "A": 1,
+    "B": 1
   }
 
   DL_EXTRACT_RESULT = {
-    "data": "data",
+    "mark_data": "data",
+    "phantom_data": "data",
     "train": "train.csv",
     "test": "test.csv",
     "validation": "validation.csv",
-    "invivo": "invivo.csv"
+    "A": "train.csv",
+    "B": "train.csv",
+    "MARK": "mark.csv"
   }
 
 if __name__ == "__main__":