+++ docs/release_notes.md
@@ -11,3 +11,5 @@
 *   Better proxy support, possibility to add certificate
 *   Add `decoders` kwargs to override the default feature decoding
     ([guide](https://github.com/tensorflow/datasets/tree/master/docs/decode.md)).
+*   Add `duke_ultrasound` dataset of ultrasound phantoms and invivo liver 
+    images from the [MimickNet paper](https://arxiv.org/abs/1908.05782)